Create Qualcomm_QCNCM865_m.2.txt
The linux driver for this chip is very actively being worked on. Last week with a 6.8 kernel, it wasn't working. Today after upgrading to a 6.9 kernel (and the latest arch firmware bundle), it is working!

I'm not quite sure how the regulatory is handled with the ath12k driver. It is self-managed, but it doesn't seem to update based on a scan like the intel chips do.